MedBotAI/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
backports.strenum==1.3.1
bcrypt==4.2.0
bidict==0.23.1
bitarray==3.0.0
blinker==1.8.2
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
databricks-sdk==0.35.0
dataclasses-json==0.6.7
datasets==3.0.2
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docopt==0.6.2
durationpy==0.9
eventlet==0.35.2
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastapi==0.115.2
filelock==3.16.1
Flask>=2.0.0
Flask-Cors>=4.0.0
Flask-Limiter==3.8.0
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
giskard==2.15.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
griffe==0.48.0
grpcio==1.67.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.1.5
langchain-community==0.3.1
langchain-core==0.3.11
langchain-huggingface==0.1.0
langchain-openai==0.0.5
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.131
lark==1.2.2
limits==3.13.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mixpanel==4.10.1
mlflow-skinny==2.17.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msal==1.31.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
num2words==0.5.13
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.2
openai>=1.0.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
ordered-set==4.1.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
posthog==3.7.0
proto-plus==1.24.0
protobuf==4.25.5
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.2.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
ragas==0.2.3
redis==5.1.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rouge_score==0.1.2
rsa==4.9
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.4
sentence-transformers==3.1.1
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.40.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
tomli==2.0.2
torch==2.5.1
torchaudio==2.5.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
zstandard==0.23.0
pathlib>=1.0.1
logging>=0.5.1.2
